Reorganized files, changed all double to float

+++ inc/ai_aircraft.h
@@ -10,14 +10,14 @@
 
 typedef struct
 {
-  double dx, dy, dz;
+  float dx, dy, dz;
   int duration;
 } Heading;
 
 typedef struct
 {
-  double x, y, z;
-  double roll, pitch, yaw;
+  float x, y, z;
+  float roll, pitch, yaw;
   int num_headings;
   Heading *headings;
   int counter;
@@ -29,7 +29,7 @@ void ai_aircraft_init(AiAircraft *aircraft, int num_aircraft);
 void ai_aircraft_update_pose(AiAircraft *aircraft, int num_aircraft, int frame_counter);
 void ai_aircraft_copy(AiAircraft *dest, const AiAircraft *src, int count);
 void ai_aircraft_move(AiAircraft *aircraft, Heading heading);
-void ai_aircraft_update_status(AiAircraft * crafts, int num_aircraft);
-void normalize_vector(double *dx, double *dy, double *dz);
+void ai_aircraft_update_status(AiAircraft *crafts, int num_aircraft);
+void normalize_vector(float *dx, float *dy, float *dz);
 
 #endif // AI_AIRCRAFT_H
