+++ src/debug.c
@@ -17,12 +17,14 @@
 /*********************************************************************************************
  * @brief Writes a string to the SWO debug output
  * @param str Null-terminated string to output
- * 
+ *
  * Outputs each character of the string through SWO channel 0
  *********************************************************************************************/
-void debug_print(const char* str) {
-  for (size_t i = 0; i < strlen(str); i++) {
-      sl_debug_swo_write_u8(0, str[i]);
+void debug_print(const char *str)
+{
+  for (size_t i = 0; i < strlen(str); i++)
+  {
+    sl_debug_swo_write_u8(0, str[i]);
   }
 }
 
@@ -30,10 +32,11 @@ void debug_print(const char* str) {
  * @brief Prints formatted text followed by a newline
  * @param format Printf-style format string
  * @param ... Variable arguments matching format string
- * 
+ *
  * Combines debug_printf() functionality with automatic newline
  *********************************************************************************************/
-void debug_println(const char *format, ...) {
+void debug_println(const char *format, ...)
+{
   va_list args;
   debug_printf(format, args);
   debug_print("\n");
@@ -44,28 +47,29 @@ void debug_println(const char *format, ...) {
  * @param buffer Output buffer to store the resulting string
  * @param size Size of the output buffer
  * @param number Floating-point number to convert
- * 
+ *
  * Converts number to string with 4 decimal places of precision
  * Handles negative numbers and includes a trailing space
  *********************************************************************************************/
-void float_to_string(char *buffer, size_t size, double number) {
+void float_to_string(char *buffer, size_t size, float number)
+{
   int integerPart = fabs((int)number);
-  double fractionalPart = fabs(number - integerPart);
+  float fractionalPart = fabs(number - integerPart);
   int fractionalInt = (int)(fractionalPart * 10000 + 0.5);
 
   snprintf(buffer, size, "%s%d.%04d ", number < 0 ? "-" : "", integerPart, fractionalInt);
-
 }
 
 /*********************************************************************************************
  * @brief Prints formatted text to SWO debug output
  * @param format Printf-style format string
  * @param ... Variable arguments matching format string
- * 
+ *
  * Provides printf-like formatting with output through SWO
  * Limited to 512 characters of output
  *********************************************************************************************/
-void debug_printf(const char *format, ...) {
+void debug_printf(const char *format, ...)
+{
   char buffer[512];
 
   va_list args;
@@ -81,12 +85,12 @@ void debug_printf(const char *format, ...) {
 /*********************************************************************************************
  * @brief Prints a floating-point number to SWO debug output
  * @param num Number to print
- * 
+ *
  * Converts the number to string and outputs it with 4 decimal places
  *********************************************************************************************/
-void debug_print_float(double num){
+void debug_print_float(float num)
+{
   char value[50];
   float_to_string(value, sizeof(value), num);
   debug_printf("%s", value);
 }
-
